+++ activerecord/test/cases/adapters/postgresql/schema_test.rb
@@ -476,7 +476,7 @@ class SchemaTest < ActiveRecord::PostgreSQLTestCase
     end
 
     def bind_param(value)
-      ActiveRecord::Relation::QueryAttribute.new(nil, value, ActiveRecord::Type::Value.new)
+      ActiveRecord::Relation::QueryAttribute.new(nil, value, ActiveModel::Type::Value.new)
     end
 end