+++ activerecord/test/cases/adapters/sqlite3/sqlite3_adapter_test.rb
@@ -82,7 +82,7 @@ module ActiveRecord
 
       def test_exec_insert
         with_example_table do
-          vals = [Relation::QueryAttribute.new("number", 10, Type::Value.new)]
+          vals = [Relation::QueryAttribute.new("number", 10, ActiveModel::Type::Value.new)]
           @conn.exec_insert("insert into ex (number) VALUES (?)", "SQL", vals)
 
           result = @conn.exec_query(
@@ -150,7 +150,7 @@ module ActiveRecord
         with_example_table "id int, data string" do
           @conn.exec_query('INSERT INTO ex (id, data) VALUES (1, "foo")')
           result = @conn.exec_query(
-            "SELECT id, data FROM ex WHERE id = ?", nil, [Relation::QueryAttribute.new(nil, 1, Type::Value.new)])
+            "SELECT id, data FROM ex WHERE id = ?", nil, [Relation::QueryAttribute.new(nil, 1, ActiveModel::Type::Value.new)])
 
           assert_equal 1, result.rows.length
           assert_equal 2, result.columns.length